refactor(wordlist): extract wordlist logic into separate component
The change improves code organization by:
- Moving wordlist functionality into dedicated WordlistObjects component
- Creating proper TypeScript interfaces for wordlist data structures
- Maintaining all existing features including pagination, search, and swipe gestures
- Simplifying the main Wordlist component to focus on layout structure
